Favorite code block
Overall, I'm proud of my entire js file - but the game start was perhaps the most satisfying. With one click, so many things happen! The name input value populates on the header, the cat age information is unhidden from the header, and the game officially starts with the initiation of all four timers.